+++ src/components/Carousel/Body.jsx
@@ -49,7 +49,7 @@ const CarouselBody = ({ data, dataBlock, isEditMode }) => {
             as={UniversalLink}
             href={href['@id']}
             target={data.openLinkInNewTab ? '_blank' : null}
-            tabindex="-1"
+            tabIndex="-1"
           >
             <>
               {(href.hasPreviewImage || image) && (