+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Bugfix
 
+- Fix `tabindex` typo @sneridagh
+
 ### Internal
 
 ## 9.4.1 (2022-01-25)