+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 9.4.2 (unreleased)
+## 9.5.0 (2022-02-08)
 
 ### Feature